Enable post edit and delete options only for the creator

+++ backend/routes/posts.js
@@ -59,10 +59,16 @@ router.put('/:id', checkAuth, multer({ storage: storage }).single('image'), (req
     content: req.body.content,
     imagePath: imagePath
   });
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    res.status(200).json({
-      message: 'Post updated successfully'
-    });
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({
+        message: 'Post updated successfully'
+      });
+    } else {
+      res.status(401).json({
+        message: 'Not authorized'
+      });
+    }
   });
 });
 
@@ -99,10 +105,16 @@ router.get('/:id', (req, res, next) => {
 })
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    res.status(200).json({
-      message: 'Post deleted successfully'
-    });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({
+        message: 'Post updated successfully'
+      });
+    } else {
+      res.status(401).json({
+        message: 'Not authorized'
+      });
+    }
   });
 });
